Add files via upload
This file has analysis of various time series data using the benchmark models Naive, Seasonal Naive, Mean and Drift. It goes through forecasting and decomposition, forecasting with models and analysing residuals to check if the assumptions associated with the residuals are met since they affect the quality of forecasts and prediction intervals. The analysis is done using R and the fable package.